# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is e-commerce server.",
        "title": "E Commerce API",
        "termsOfService": "https://e-ticaret-api.herokuapp.com/api/v1/",
        "contact": {
            "name": "Api Support",
            "url": "https://e-ticaret-api.herokuapp.com/support",
            "email": "suatcnby06@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:3500",
    "basePath": "/api/v1",
    "paths": {
...
        "/comments": {
            "get": {
                "description": "Get all comments",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Comments"
                ],
                "summary": "All  Data",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Offset",
                        "name": "offset",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.Comment"
                        }
                    }
                }
            },
            "post": {
                "description": "create comments",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Comments"
                ],
                "summary": "Create Data",
                "parameters": [
                    {
                        "description": "comment form",
                        "name": "body",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/entity.Comment"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/entity.Comment"
                            }
                        }
                    }
                }
            }
        },
        "/comments/{id}": {
            "get": {
                "description": "get string by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Comments"
                ],
                "summary": "Show Data",
                "parameters": [
                    {
                        "type": "string",
                        "description": "comment ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.Comment"
                        }
                    }
                }
            },
            "put": {
                "description": "update comment",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Comments"
                ],
                "summary": "Update Data",
                "parameters": [
                    {
                        "type": "string",
                        "description": "comment ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "comment update form",
                        "name": "body",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/entity.Comment"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.Comment"
                        }
                    }
                }
            },
            "delete": {
                "description": "delete comments",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Comments"
                ],
                "summary": "Delete Data",
                "parameters": [
                    {
                        "type": "string",
                        "description": "comment ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/entity.Comment"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "entity.Comment": {
            "type": "object",
            "required": [
                "comment",
                "ip",
                "rate",
                "status"
            ],
            "properties": {
                "comment": {
                    "type": "string"
                },
                "contentId": {
                    "type": "integer"
                },
                "ip": {
                    "type": "string"
                },
                "rate": {
                    "type": "number"
                },
                "status": {
                    "type": "boolean"
                },
                "userId": {
                    "type": "integer"
                }
            }
        },
...
    }
}